Extract getByIdOrThrow
